[Game] updated Transfer - add transfer_path.xml - add TransferMoveType - fix UnitMoveTipe [Game] updated Simulation - add Transfer [Game] add rotation in CharTemplates.json [Game] add rotation in Worlds.json [Game] rename Enum SkillCasterType to EffectOriginType - update enum [Game] update SCUnitState [Game] add enums AttachPoint, AttachUnitReason [Game] moved LERP function to MathUtil.cs [Game] update some Packets [Game] added Enums AttachPoint and ModelKind

+++ AAEmu.Game/Models/Game/Units/AttachUnitReason.cs
@@ -1,6 +1,6 @@
 ﻿namespace AAEmu.Game.Models.Game.Units
 {
-    public enum AttachUnitReason : int
+    public enum AttachUnitReason : byte
     {
         None = 0,
         MountMateLeft = 1,
